+++ packages/react/src/useMount.ts
@@ -1,0 +1,6 @@
+import ComponentSetup from './ComponentSetup';
+
+export default function useMount(fn: () => void) {
+  const setup = ComponentSetup.getCurrent<ComponentSetup>();
+  setup.mountFns.push(fn);
+}